// src/store/modules/auth/enums.ts
export enum Tokens {
  STORE_KEY_TOKEN = "token",
  STORE_KEY_REFRESH_TOKEN = "refresh-token"
}

export enum MutationTypes {
  SET_TOKEN = "AUTH_SET_TOKEN",
  SET_LOGOUT = "AUTH_SET_LOGOUT",
  SET_REFRESH_TOKEN = "AUTH_SET_REFRESH_TOKEN",
  SET_USER_INFO = "AUTH_SET_USER_INFO"
}

export enum ActionTypes {
  AJAX_CHECK_AUTHENTICATED = "AUTH_AJAX_CHECK_AUTHENTICATED",
  AJAX_LOGIN = "AUTH_AJAX_LOGIN",
  AJAX_LOGOUT = "AUTH_AJAX_LOGOUT",
  AJAX_REFRESH_TOKEN = "AUTH_AJAX_REFRESH_TOKEN",
  AJAX_USER_INFO = "AUTH_AJAX_USER_INFO",
  AJAX_CHNAGE_PASSWORD = "AUTH_AJAX_CHNAGE_PASSWORD"
}

export enum GetterTypes {
  GET_AUTHENTICATED = "AUTH_GET_AUTHENTICATED",
  GET_TOKEN = "AUTH_GET_TOKEN",
  GET_REFRESH_TOKEN = "AUTH_GET_REFRESH_TOKEN",
  GET_USER_INFO = "AUTH_GET_USER_INFO"
}